backend/pyproject.toml
[project]
name = "novel-agent"
version = "0.1.0"
description = "Novelist AI backend based on short-movie-agents architecture."
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "google-adk~=1.14.0",
    "fastapi~=0.115.8",
    "uvicorn~=0.34.0",
    "google-cloud-logging>=3.12.0",
    "opentelemetry-exporter-gcp-trace~=1.9.0",
    "sqlmodel>=0.0.31", 
    "python-dotenv>=1.2.1",
    "google-generativeai>=0.8.3",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"